requirements.txt
numpy
pandas~=2.2.3
faker
argparse
gspread~=6.1.3
google-auth